website: present provider documentation via categories
We have outgrown the single flat list presentation of providers
due to the sheer number now present, so we'll move here to a
model where the providers are split into a number of categories
that each contain a smaller list.

The full list is still included in the body of the main index
page for quick access via search, but the categories make for
a more accessible navbar for those who are just browsing.